# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Técnico - UPD8",
        "description": "Desafio Técnico - UPD8",
        "contact": {
            "email": "acarlos.os@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://desafioupd8.teste/api",
            "description": "local"
        }
    ],
    "paths": {
        "/cidade": {
            "get": {
                "tags": [
                    "Cidade"
                ],
                "summary": "Return cidades",
                "description": "Return cidades",
                "operationId": "cidade-index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "types": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "nome": {
                                                                "type": "string",
                                                                "example": "Acrelândia"
                                                            },
                                                            "uf": {
                                                                "type": "string",
                                                                "example": "AC"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-07-09T22:22:07.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2024-07-09T22:22:07.000000Z"
                                                            },
                                                            "deleted_at": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cidade"
                ],
                "summary": "Create cidade",
                "description": "Create cidade",
                "operationId": "cidade-store",
                "requestBody": {
                    "description": "Create cidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "uf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Campinas"
                                    },
                                    "uf": {
                                        "type": "string",
                                        "example": "SP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Campinas"
                                        },
                                        "uf": {
                                            "type": "string",
                                            "example": "SP"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-07-09T22:22:07.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-07-09T22:22:07.000000Z"
                                        },
                                        "deleted_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Wrong error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong error. Please try again!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}